musicmarketplace/musicmarketplace_backend/requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
